started working on new Data system to reduce runtime panics
the old custom_sets system was very flexible, but required the user to do a lot
of work on their own (creating the sets) and was a very finnicky system.
Hopefully the new system is easier to use, safer, and can provide better error
messages when it does go wrong.